рефакторинг: распределение классов по файлам
